# DuckType extensions optimisation