This PR adds a cache layer for `Type.IsPublic` and `Type.IsNestedPublic` to avoid the extern call on `TryDuckCast`, `DuckAs` and `DuckIs` extension methods.

### Benchmarks:

#### Current:
```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.928 (2004/?/20H1)
Intel Core i7-1068NG7 CPU 2.30GHz, 1 CPU, 2 logical and 2 physical cores
.NET Core SDK=5.0.202
  [Host]     : .NET Core 3.1.14 (CoreCLR 4.700.21.16201, CoreFX 4.700.21.16208), X64 RyuJIT
  DefaultJob : .NET Core 3.1.14 (CoreCLR 4.700.21.16201, CoreFX 4.700.21.16208), X64 RyuJIT


|      Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|---------:|---------:|-------:|------:|------:|----------:|
| TryDuckCast | 29.93 ns | 0.254 ns | 0.225 ns | 0.0025 |     - |     - |      24 B |
|    DuckCast | 22.92 ns | 0.406 ns | 0.380 ns | 0.0025 |     - |     - |      24 B |
|      DuckAs | 30.03 ns | 0.169 ns | 0.158 ns | 0.0024 |     - |     - |      24 B |
|      DuckIs | 18.91 ns | 0.147 ns | 0.123 ns |      - |     - |     - |         - |
```

#### PR:
```

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.928 (2004/?/20H1)
Intel Core i7-1068NG7 CPU 2.30GHz, 1 CPU, 2 logical and 2 physical cores
.NET Core SDK=5.0.202
  [Host]     : .NET Core 3.1.14 (CoreCLR 4.700.21.16201, CoreFX 4.700.21.16208), X64 RyuJIT
  DefaultJob : .NET Core 3.1.14 (CoreCLR 4.700.21.16201, CoreFX 4.700.21.16208), X64 RyuJIT


|      Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|---------:|---------:|-------:|------:|------:|----------:|
| TryDuckCast | 20.64 ns | 0.303 ns | 0.253 ns | 0.0025 |     - |     - |      24 B |
|    DuckCast | 23.05 ns | 0.486 ns | 0.681 ns | 0.0025 |     - |     - |      24 B |
|      DuckAs | 23.52 ns | 0.379 ns | 0.601 ns | 0.0025 |     - |     - |      24 B |
|      DuckIs | 11.05 ns | 0.028 ns | 0.023 ns |      - |     - |     - |         - |
```


@DataDog/apm-dotnet